requirements.txt
fastapi
pydantic
uvicorn
pandas
python-multipart
openpyxl
xlwt
requests
pyyaml